backend/requirements.txt
Flask
Flask_SQLAlchemy
flask-cors
psycopg2-binary
python-dotenv